pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Willem <1121567132@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.15.1"
pillow = "^9.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# change pypi source
[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
default = true